rating/requirements.txt
redis
psycopg2-binary
sqlalchemy
asyncio
python-dotenv
loguru 